+++ src/Puphpet/MainBundle/Resources/views/manifest/puphpet/shell/install-ruby.sh
@@ -17,8 +17,13 @@ echo 'Installing Ruby 1.9.3 using RVM'
 curl -sSL https://get.rvm.io | bash -s stable --ruby=1.9.3
 source /usr/local/rvm/scripts/rvm
 
-mv /usr/bin/ruby /usr/bin/ruby-old
-mv /usr/bin/gem /usr/bin/gem-old
+if [[ -f '/usr/bin/ruby' ]]; then
+    mv /usr/bin/ruby /usr/bin/ruby-old
+fi
+
+if [[ -f '/usr/bin/gem' ]]; then
+    mv /usr/bin/gem /usr/bin/gem-old
+fi
 
 ln -s /usr/local/rvm/rubies/ruby-1.9.3-*/bin/ruby /usr/bin/ruby
 ln -s /usr/local/rvm/rubies/ruby-1.9.3-*/bin/gem /usr/bin/gem
